sentiment-project/requirements.txt
textblob==0.17.1
nltk==3.8.1
requests==2.31.0
beautifulsoup4==4.12.2
ollama>=0.1.0
python-dotenv>=1.0.0
